Refactor: Changed Folder Structure
