+++ src/Utils/AppReducer.js
@@ -3,7 +3,7 @@ export const initialState = {
   user: null
 }
 
-export const reducer = (state, action) => {
+export const AppReducer = (state, action) => {
   switch (action.type) {
     case 'SET_USER':
       return { ...state, user: action.payload }
