+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from "react-router-dom"
 import App from './App';
-import { AppProvider } from "./contexts/AppContext"
+import { AppProvider } from "./Contexts/AppContext"
 
 ReactDOM.render(
   <React.StrictMode>
