+++ src/pages/LoginSignup/Login.js
@@ -1,11 +1,10 @@
-import './Styles/Login.css'
+import './LoginSignup.css'
 import { useState } from 'react'
 import { Link, useNavigate } from 'react-router-dom'
 import axios from 'axios'
-import { useApp } from '../contexts/AppContext'
+import { useApp } from '../../Contexts/AppContext'
 
-export const Signup = () => {
-  const [name, setName] = useState(null)
+export const Login = () => {
   const [email, setEmail] = useState(null)
   const [password, setPassword] = useState(null)
   const [loading, setLoading] = useState(false)
@@ -13,20 +12,17 @@ export const Signup = () => {
   const { app, dispatch } = useApp()
   const navigate = useNavigate()
 
-  const signupHandler = async () => {
+  const loginHandler = async () => {
     if (!app.loggedInToken) {
-      console.log('triggered signup')
       setLoading(true)
       try {
         const { data } = await axios.post(
-          'https://cyanic-api.herokuapp.com/signup',
+          'https://cyanic-api.herokuapp.com/login',
           {
-            name: name,
             email: email,
             password: password,
           },
         )
-        console.log(data)
         if (data.success) {
           dispatch({ type: 'SET_LOGGED_IN_TOKEN', payload: data.token })
           dispatch({ type: 'SET_USER', payload: data.user })
@@ -37,7 +33,9 @@ export const Signup = () => {
           navigate('/')
         }
         setError(data.message)
+        setLoading(false)
       } catch (error) {
+        setLoading(false)
         console.log(error)
       }
     }
@@ -50,14 +48,7 @@ export const Signup = () => {
         </div>
       ) : (
         <div>
-          <h2>Signup</h2>
-          <input
-            className="input"
-            type="text"
-            placeholder="Name"
-            onChange={(e) => setName(e.target.value)}
-          />
-          <br />
+          <h2>Login</h2>
           <input
             className="input"
             type="text"
@@ -67,20 +58,24 @@ export const Signup = () => {
           <br />
           <input
             className="input"
-            type="text"
+            type="password"
             placeholder="Password"
             onChange={(e) => setPassword(e.target.value)}
           />
           <br />
-          <button className="btn btnPrimary" onClick={() => signupHandler()}>
-            {loading ? 'Signing Up...' : 'Signup'}
+          <button
+            type="submit"
+            className="btn btnPrimary"
+            onClick={loginHandler}
+          >
+            {loading ? 'Logging In...' : 'Login'}
           </button>
         </div>
       )}
       <p>
-        <small>Already a user? </small>
-        <Link to="/login">
-          <bold>Login</bold>
+        <small>Not a user? </small>
+        <Link to="/signup">
+          <bold>Signup</bold>
         </Link>
       </p>
       {error && <h3>Some error occured, try again.</h3>}
