+++ src/pages/VideoPlayer/VideoPlayer.js
@@ -1,7 +1,7 @@
-import './Styles/VideoPlayer.css'
+import './VideoPlayer.css'
 import { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
-import { useApp } from '../contexts/AppContext'
+import { useApp } from '../../Contexts/AppContext'
 import axios from 'axios'
 import { useNavigate } from 'react-router-dom'
 import { FaRegShareSquare } from 'react-icons/fa'
