+++ src/pages/index.js
@@ -1,8 +1,8 @@
-export { HomePage } from './HomePage'
-export { LikedVideos } from './LikedVideos'
-export { Playlists } from './Playlists'
-export { PlaylistDetail } from './PlaylistDetail'
-export { VideoPlayer } from './VideoPlayer'
-export { WatchLater } from './WatchLater'
-export { Login } from './Login'
-export { Signup } from './Signup'
+export { HomePage } from './HomePage/HomePage'
+export { LikedVideos } from './LikedVideos/LikedVideos'
+export { Playlists } from './Playlists/Playlists'
+export { PlaylistDetail } from './PlaylistDetail/PlaylistDetail'
+export { VideoPlayer } from './VideoPlayer/VideoPlayer'
+export { WatchLater } from './WatchLater/WatchLater'
+export { Login } from './LoginSignup/Login'
+export { Signup } from './LoginSignup/Signup'
